# api/openapi.yaml
openapi: 3.0.3
info:
  title: IMS ERP System API
  description: |
    Complete ERP system API with microservices architecture
    
    ## Architecture
    - **API Gateway** (8080) - Central entry point for all client requests
    - **Auth Service** (8081) - Authentication and authorization
    - **Client Services** (8082) - Client management (CQRS pattern)
    - **Invoice Service** (8083) - Invoice management
    - **Payment Service** (8084) - Payment processing
    - **Product Service** (8085) - Product catalog
    - **Order Service** (8086) - Order management
    - **Inventory Service** (8087) - Inventory control
    
    ## Authentication
    All protected endpoints require Bearer token authentication.
    Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-token>
    ```
    
    ## Rate Limiting
    API requests are rate limited to 1000 requests per minute per tenant.
    
    ## Error Handling
    All errors follow standard HTTP status codes with detailed error messages.
  version: 1.0.0
  contact:
    name: IMS ERP Support
    email: support@ims-erp.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.ims-erp.example.com/api/v1
    description: Production server
...
paths:
...
  /payments/process:
    post:
      tags:
        - Payments
      summary: Process payment
      description: Processes a payment through the specified provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessPaymentRequest'
      responses:
        '201':
          description: Payment processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResult'
  /payments/refund:
    post:
      tags:
        - Payments
      summary: Process refund
      description: Processes a refund for a previous payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundRequest'
      responses:
        '201':
          description: Refund processed successfully
...
components:
  schemas:
...
    ProcessPaymentRequest:
      type: object
      required:
        - invoiceId
        - amount
        - currency
        - method
      properties:
        invoiceId:
          type: string
          format: uuid
        amount:
          type: number
        currency:
          type: string
        method:
          type: string
          enum: [credit_card, bank_transfer, paypal, stripe]
        provider:
          type: string
        description:
          type: string
    PaymentResult:
      type: object
      properties:
        success:
          type: boolean
        paymentId:
          type: string
        transactionId:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed]
    RefundRequest:
      type: object
      required:
        - paymentId
        - amount
        - reason
      properties:
        paymentId:
          type: string
          format: uuid
        amount:
          type: number
        reason:
          type: string
        refundType:
          type: string
          enum: [full, partial]
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []